requirements.txt
Flask==3.0.0
requests==2.31.0
python-pptx==0.6.23
cos-python-sdk-v5==1.9.25
Werkzeug==3.0.1